requirements.txt
openstacksdk
dotenv